pyproject.toml
[project]
name = "langchain-ppio"
version = "0.0.1"
description = "PPIO sandbox integration for Deep Agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "deepagents>=0.4.4",
    "e2b-code-interpreter>=2.4.1",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://docs.langchain.com/oss/python/deepagents/sandboxes"
Repository = "https://github.com/langchain-ai/deepagents"
Documentation = "https://docs.langchain.com/oss/python/deepagents/sandboxes"

[project.scripts]
langchain-ppio-smoke = "langchain_ppio.smoke:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langchain_ppio"]

[tool.hatch.build.targets.sdist]
include = [
    "src/langchain_ppio",
    "tests",
    "README.md",
    "LICENSE",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
